+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix URLs handling (no double slashes) [#2](https://github.com/opendatateam/udata-ods/pull/2)
 - Fix packaging [#3](https://github.com/opendatateam/udata-ods/pull/3)
 - Make use of [udata pytest plugin](opendatateam/udata#1400) [#4](https://github.com/opendatateam/udata-ods/pull/4)
+- Alternative exports support [#14](https://github.com/opendatateam/udata-ods/pull/14)
 
 ## 1.0.0 (2017-10-20)